AND-6831 Renamed UserWalletListManager package.
