+++ app/src/main/java/com/tangem/tap/domain/userWalletList/implementation/BiometricUserWalletsListManager.kt
@@ -1,9 +1,9 @@
 package com.tangem.tap.domain.userWalletList.implementation
 
 import com.tangem.common.*
-import com.tangem.domain.wallets.legacy.UserWalletsListError
-import com.tangem.domain.wallets.legacy.UserWalletsListManager
-import com.tangem.domain.wallets.legacy.UserWalletsListManager.Lockable.UnlockType
+import com.tangem.domain.wallets.manager.UserWalletsListError
+import com.tangem.domain.wallets.manager.UserWalletsListManager
+import com.tangem.domain.wallets.manager.UserWalletsListManager.Lockable.UnlockType
 import com.tangem.domain.wallets.models.UserWallet
 import com.tangem.domain.wallets.models.UserWalletId
 import com.tangem.tap.domain.userWalletList.model.UserWalletEncryptionKey
