+++ app/src/main/java/com/tangem/tap/features/welcome/ui/WelcomeViewModel.kt
@@ -4,7 +4,7 @@ import android.content.Intent
 import androidx.lifecycle.*
 import com.tangem.common.core.TangemError
 import com.tangem.core.analytics.Analytics
-import com.tangem.domain.wallets.legacy.UserWalletsListError
+import com.tangem.domain.wallets.manager.UserWalletsListError
 import com.tangem.tap.common.analytics.events.SignIn
 import com.tangem.tap.common.redux.global.GlobalAction
 import com.tangem.tap.features.details.ui.cardsettings.TextReference
