+++ domain/wallets/src/main/java/com/tangem/domain/wallets/usecase/GetWalletsUseCase.kt
@@ -1,6 +1,6 @@
 package com.tangem.domain.wallets.usecase
 
-import com.tangem.domain.wallets.legacy.UserWalletsListManager
+import com.tangem.domain.wallets.manager.UserWalletsListManager
 import com.tangem.domain.wallets.models.UserWallet
 import kotlinx.coroutines.flow.Flow
