+++ domain/wallets/src/main/java/com/tangem/domain/wallets/usecase/WalletNameGenerateUseCase.kt
@@ -1,7 +1,7 @@
 package com.tangem.domain.wallets.usecase
 
 import com.tangem.domain.models.scan.ProductType
-import com.tangem.domain.wallets.legacy.UserWalletsListManager
+import com.tangem.domain.wallets.manager.UserWalletsListManager
 
 /**
  * Use case for user wallet name generation
